Renamed CommonsTypesGenerators to CommonGraphGenerators
