+++ triples-generator/src/test/scala/ch/datascience/triplesgenerator/generators/ServiceTypesGenerators.scala
@@ -19,7 +19,7 @@
 package ch.datascience.triplesgenerator.generators
 
 import ch.datascience.generators.Generators._
-import ch.datascience.generators.CommonsTypesGenerators._
+import ch.datascience.generators.CommonGraphGenerators._
 import ch.datascience.triplesgenerator.config.DatasetType.{Mem, TDB}
 import ch.datascience.triplesgenerator.config._
 import ch.datascience.triplesgenerator.eventprocessing.RDFTriples
