+++ webhook-service/test/ch/datascience/webhookservice/eventprocessing/pushevent/IOCommitInfoFinderSpec.scala
@@ -25,7 +25,7 @@ import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators.exceptions
 import ch.datascience.graph.model.events.CommittedDate
 import ch.datascience.graph.model.events.EventsGenerators._
-import ch.datascience.generators.CommonsTypesGenerators._
+import ch.datascience.generators.CommonGraphGenerators._
 import ch.datascience.stubbing.ExternalServiceStubbing
 import ch.datascience.webhookservice.config.GitLabConfig._
 import ch.datascience.webhookservice.config.IOGitLabConfigProvider
