+++ webhook-service/test/ch/datascience/webhookservice/hookcreation/EventsHistoryLoaderSpec.scala
@@ -23,7 +23,7 @@ import cats.implicits._
 import ch.datascience.http.client.AccessToken
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators._
-import ch.datascience.generators.CommonsTypesGenerators._
+import ch.datascience.generators.CommonGraphGenerators._
 import ch.datascience.graph.model.events._
 import ch.datascience.interpreters.TestLogger
 import ch.datascience.interpreters.TestLogger.Level.{Error, Info}
